Add files via upload
Marriage_Market_Oct_21_short relations is a non-spatial agent-based model of the formation of romantic dyads.  It was developed using the Numerus software (https://www.numerusinc.com/) by Nancy Darling and Ian Burns.  Numerus was developed by Richard Salter. 

The model was designed to examine sampling bias in cross-sectional samples of romantic relationships. We assumed three relationship styles, loosely based on attachment theory: secure, avoidant, and anxious.  Each style differs in mean relationship duration.  Couples are randomly paired, with the duration of the new relationship assigned as the mean of the durations of the two styles.